# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bus Booking System",
        "version": "0.1"
    },
    "paths": {
...
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "description": "retrieve authed user bookings",
                "operationId": "4535e967ef15bc76dadf1efa40192f0d",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "If any error happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Not Authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/bookings/{id}": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "description": "retrieve authed user selected booking ",
                "operationId": "c383a28985575aa1a16f783b2f762554",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of booking",
                        "required": true,
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "If any error happened",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "Not Authenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/api/bookings/book": {
            "post": {
                "tags": [
                    "Bookings"
                ],
                "description": "book seat",
                "operationId": "6a7cdf33f70e0ebc233a261cc3ca4481",
                "requestBody": {
                    "description": "booking paramters (trip_id ,seat_id ,  source_station_id , destination_station_id  )",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "trip_id",
                                    "seat_id",
                                    "source",
                                    "destination"
                                ],
                                "properties": {
                                    "trip_id": {
                                        "type": "interger",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "seat_id": {
                                        "type": "interger",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "source_station_id": {
                                        "type": "interger",
                                        "format": "text",
                                        "example": "1"
                                    },
                                    "destination_station_id": {
                                        "type": "interger",
                                        "format": "text",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Booking stored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "description": "Success"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "format": "text",
                                            "example": "You Are Not Authorized!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Fill Data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}